Refactored uploaders
